bdevperf: change the type of time_in_sec to int.
Since we use time_in_sec to judge < 0 case

Reviewed-on: https://review.gerrithub.io/391154
Tested-by: SPDK Automated Test System <sys_sgsw@intel.com>

+++ test/lib/bdev/bdevperf/bdevperf.c
@@ -714,7 +714,7 @@ main(int argc, char **argv)
 	int op;
 	bool mix_specified;
 	struct spdk_app_opts opts = {};
-	uint64_t time_in_sec;
+	int time_in_sec;
 
 	/* default value */
 	config_file = NULL;
